Account creation

+++ src/main/java/com/jyqdesign/keysbankbackend/repository/extractor/AccountResultSetExtractor.java
@@ -27,21 +27,21 @@ public class AccountResultSetExtractor implements ResultSetExtractor<List<Accoun
                 account.setName(rs.getString("account_name"));
                 account.setBank(rs.getString("account_bank"));
                 account.setReference(rs.getString("account_reference"));
-                account.setCurrentSolde(rs.getInt("account_current_solde"));
-                account.setInitialBalance(rs.getInt("account_initial_balance"));
+                account.setCurrentSolde(rs.getLong("account_current_solde"));
+                account.setInitialBalance(rs.getLong("account_initial_balance"));
                 account.setLastUpdate(rs.getString("account_last_update"));
-                account.setSinceYear(rs.getInt("account_since_year"));
+                account.setSinceYear(rs.getLong("account_since_year"));
                 account.setCsvImportFolder(rs.getString("csv_import_folder"));
-                account.setCsvRowIgnored(rs.getInt("csv_row_ignored"));
-                account.setCsvRowDate(rs.getInt("csv_row_date"));
-                account.setCsvRowSolde(rs.getInt("csv_row_solde"));
-                account.setCsvColumnNumber(rs.getInt("csv_column_number"));
-                account.setCsvColumnDate(rs.getInt("csv_column_date"));
+                account.setCsvRowIgnored(rs.getLong("csv_row_ignored"));
+                account.setCsvRowDate(rs.getLong("csv_row_date"));
+                account.setCsvRowSolde(rs.getLong("csv_row_solde"));
+                account.setCsvColumnNumber(rs.getLong("csv_column_number"));
+                account.setCsvColumnDate(rs.getLong("csv_column_date"));
                 account.setCsvDateFormat(rs.getString("csv_date_format"));
-                account.setCsvColumnDescription(rs.getInt("csv_column_description"));
-                account.setCsvColumnCredit(rs.getInt("csv_column_credit"));
-                account.setCsvColumnDebit(rs.getInt("csv_column_debit"));
-                account.setCsvColumnValue(rs.getInt("csv_column_value"));
+                account.setCsvColumnDescription(rs.getLong("csv_column_description"));
+                account.setCsvColumnCredit(rs.getLong("csv_column_credit"));
+                account.setCsvColumnDebit(rs.getLong("csv_column_debit"));
+                account.setCsvColumnValue(rs.getLong("csv_column_value"));
 
                 accountMap.put(accountId, account);
             }
